# [DCA] [orchestrator] optimize scrubbing for orchestrator containers