### What does this PR do?

Testing a new scrubbing algorithm for the orchestrator related containers. 
With this change we try to support a smaller subset of supported scrubbing which would reduce the load on the agent. 
This uses literalFilter instead of regexp matches. Which should, from my gut feeling, catch ~80% of the usages.
For the other 20% usages we could add the possibility to supply custom regex which we parse and supply like for process-agent.

This PR shows the possibility of adding Regexes which will be compiled and matched against in every run before using literal matches.

So one can add a custom regex which will be run before literal matching. With this we can ensure that we match things we want in case literal matching does not match everything.

#### Supported:
- `*<word>*`
- `word`

#### Not Supported:
- `*word`
- `word*`
- `any kind of regexp (wo???d)`

#### Problems with the proposed solution:
does pseudo match grouping by a simple algorithm used by the orchestrator. If a word is found (contains) the next word will be replaced OR if the current word contains a "=" the word after "=" will be replaced. This can be quite error prone, hence this should be a place for discussion.

### Motivation

Reduce backtracking of regex. Which takes roughly 10% of agent time on the DCA (depending on the kind of workload the DCA is working with because we uses caches)

This is roughly 5x faster.

Benchmarking for the whole Container scrubbing
```
BenchmarkRegexMatching1000
BenchmarkRegexMatching1000/simplified
BenchmarkRegexMatching1000/simplified-8         	      30	  38323824 ns/op
BenchmarkRegexMatching1000/default
BenchmarkRegexMatching1000/default-8            	       7	 158356520 ns/op
```

Benchmark container Scrubbing with additional regex added (2 regex)
```
BenchmarkRegexMatchingCustom1000
BenchmarkRegexMatchingCustom1000/simplified
BenchmarkRegexMatchingCustom1000/simplified-8         	      79	  15146546 ns/op
BenchmarkRegexMatchingCustom1000/default
BenchmarkRegexMatchingCustom1000/default-8            	      12	  94521672 ns/op
```

Benchmarking for Command parsing only
```
BenchmarkCommandMatching1000
BenchmarkCommandMatching1000/simplified
BenchmarkCommandMatching1000/simplified-8         	      39	  29245263 ns/op
BenchmarkCommandMatching1000/default
BenchmarkCommandMatching1000/default-8            	       7	 142869598 ns/op
```

### Future
potentially we could add that kind of logic for adding additional regex in the future:

```
if canBeSimplified(word):
  s.LiteralSensitivePatterns = append(ds.LiteralSensitivePatterns, word...)
else:
  r := config.CompileStringsToRegex(words)
  ds.RegexSensitivePatterns = append(ds.RegexSensitivePatterns, r...)
```
and can be simplified checks whether the given possible regex (*.foo.*|.*bar.*) can be expressed as a literal or a concat of literals 

e.g.:

`(*.foo.*|.*bar.*) -> command.Contains(foo) || command.Contains(bar)`


### Note
We try to get this into 1.10, if we don't make it we can postpone this PR for a later release.