@sunhay 2020-10-26
Perhaps I missed the code thats responsible for it but are you caching the scrubbing at all? We had a number of performance issues with processes the first time we implemented the scrubber there, but some straightforward caching helped reduce that impact quite a bit.