@nammn 2020-10-26
> Perhaps I missed the code thats responsible for it but are you caching the scrubbing at all? We had a number of performance issues with processes the first time we implemented the scrubber there, but some straightforward caching helped reduce that impact quite a bit.

@sunhay 
I might need to note that the whole PR can more be seen as a discussion for a possible alternative scrubbing algorithm.   

regarding caching: 
- Indirectly, we cache each resource we find for I think 3 minutes, the key is the uid of the resource.
- Directly for scrubbing? No currently not, this was something I planned to add as well, maybe as a later iteration 🤔 . Finding a proper key is difficult for our case.
Potentially one could use the same `UID` cache with a higher TTL. 
Alternatively, we could try some other combinations.